+++ SimuladorNotas/nbproject/project.properties
@@ -73,7 +73,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=SimuladorNotas
-main.class=view.LoginView
+main.class=app.view.LoginView
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
